Pipeline Improvements (#45)
* Android SDK updated to latest Clang 14
* Android Pipelines now compile codegen_runtime tests
* Improved vcpkg and msvc ci scripts
* Sample now also builds codegen_runtime tests

Cmake Integration improvements
* use a better add_stencil_library method
* use find_packages for fmt and other dependencies
* Interface targets for embedding and stencil runtime
* Use FindModules for submodule vs vcpkg detection

Bugfixes and Cross Platform improvements

+++ include/stencil/transactions.binserdes.h
@@ -26,7 +26,7 @@ struct OStrmWriter
 
     auto& operator<<(shared_string const& val)
     {
-        *this << val.size();
+        *this << static_cast<uint32_t>(val.size());
         if (val.size() > 0) _ostr.write(reinterpret_cast<char const*>(val.data()), static_cast<std::streamsize>(val.size()));
         return *this;
     }
@@ -52,7 +52,7 @@ struct IStrmReader
 
     shared_string read_shared_string()
     {
-        size_t      size = read<size_t>();
+        size_t      size = read<uint32_t>();
         std::string str(size, 0);
         _istrm.read(reinterpret_cast<char*>(str.data()), static_cast<std::streamsize>(size));
         return shared_string::make(std::move(str));
